update sqlhandler async
